requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==22.0.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
